src/weatherApi/requirements.txt
django
djangorestframework
drf-spectacular
asyncio
aiohttp
django-redis
requests
psycopg2